requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
idna==3.3
multidict==5.2.0
pycparser==2.20
PyNaCl==1.4.0
PyYAML==6.0
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
youtube-dl==2021.6.6
